+++ src/BigCommerce/Plugin.php
@@ -4,7 +4,7 @@
 namespace BigCommerce;
 
 class Plugin {
-	const VERSION = '3.5.0';
+	const VERSION = '3.6.0';
 
 	protected static $_instance;