+++ src/BigCommerce/Settings/Sections/Currency.php
@@ -4,10 +4,16 @@
 namespace BigCommerce\Settings\Sections;
 
 
+use BigCommerce\Exceptions\Channel_Not_Found_Exception;
 use BigCommerce\Settings\Screens\Settings_Screen;
+use BigCommerce\Taxonomies\Channel;
 
 class Currency extends Settings_Section {
 	const NAME = 'currency';
+
+	const ENABLED_CURRENCIES = 'bigcommerce_enabled_currencies';
+
+	const CHANNEL_CURRENCY_CODE = 'bigcommerce_channel_currency_code';
 
 	const CURRENCY_CODE            = 'bigcommerce_currency_code';
 	const CURRENCY_SYMBOL          = 'bigcommerce_currency_symbol';
@@ -36,80 +42,54 @@ class Currency extends Settings_Section {
 			Settings_Screen::NAME
 		);
 
-		register_setting(
-			Settings_Screen::NAME,
-			self::CURRENCY_SYMBOL
-		);
+		try {
+			// register a separate currency field for each connected channel
+			$connections   = new Channel\Connections();
+			$active        = $connections->active();
+			$channel_count = count( $active );
+			foreach ( $active as $channel ) {
+				add_settings_field(
+					self::CHANNEL_CURRENCY_CODE . '-' . $channel->term_id,
+					$channel_count === 1 ? __( 'Currency', 'bigcommerce' ) : esc_html( $channel->name ),
+					[ $this, 'render_currency_select' ],
+					Settings_Screen::NAME,
+					self::NAME,
+					[
+						'channel'   => $channel->term_id,
+						'count'     => $channel_count,
+						'label_for' => 'field-' . self::CHANNEL_CURRENCY_CODE . '-' . $channel->term_id,
+					]
+				);
+			}
 
-		register_setting(
-			Settings_Screen::NAME,
-			self::CURRENCY_SYMBOL_POSITION
-		);
+			// we'll handle the save collectively
+			register_setting(
+				Settings_Screen::NAME,
+				self::CHANNEL_CURRENCY_CODE,
+				[
+					'sanitize_callback' => [ $this, 'save_channel_currencies' ],
+				]
+			);
+		} catch ( Channel_Not_Found_Exception $e ) {
+			add_settings_field(
+				self::CHANNEL_CURRENCY_CODE,
+				__( 'Currency', 'bigcommerce' ),
+				[ $this, 'render_default_currency_symbol' ],
+				Settings_Screen::NAME,
+				self::NAME
+			);
+		}
 
-		register_setting(
-			Settings_Screen::NAME,
-			self::DECIMAL_UNITS
-		);
 
 		register_setting(
 			Settings_Screen::NAME,
 			self::PRICE_DISPLAY
 		);
 
-
-		/**
-		 * This filter is documented in src/BigCommerce/Container/Currency.php
-		 */
-		$auto_format = apply_filters( 'bigcommerce/settings/currency/auto-format', class_exists( '\NumberFormatter' ) );
-
-		if ( ! $auto_format ) {
-			// only render currency fields if auto-format is disabled
-
-			add_settings_field(
-				self::CURRENCY_SYMBOL,
-				esc_html( __( 'Currency Symbol', 'bigcommerce' ) ),
-				[ $this, 'render_field', ],
-				Settings_Screen::NAME,
-				self::NAME,
-				[
-					'option'    => self::CURRENCY_SYMBOL,
-					'type'      => 'text',
-					'default'   => '$',
-					'label_for' => 'field-' . self::CURRENCY_SYMBOL,
-				]
-			);
-
-			add_settings_field(
-				self::CURRENCY_SYMBOL_POSITION,
-				esc_html( __( 'Symbol Position', 'bigcommerce' ) ),
-				[ $this, 'render_position_select', ],
-				Settings_Screen::NAME,
-				self::NAME,
-				[
-					'label_for' => 'field-' . self::CURRENCY_SYMBOL_POSITION,
-				]
-			);
-
-			add_settings_field(
-				self::DECIMAL_UNITS,
-				esc_html( __( 'Decimal Units', 'bigcommerce' ) ),
-				[ $this, 'render_field', ],
-				Settings_Screen::NAME,
-				self::NAME,
-				[
-					'option'    => self::DECIMAL_UNITS,
-					'type'      => 'number',
-					'default'   => 2,
-					'label_for' => 'field-' . self::DECIMAL_UNITS,
-				]
-			);
-
-		}
-
 		add_settings_field(
 			self::PRICE_DISPLAY,
 			esc_html( __( 'Price Display', 'bigcommerce' ) ),
-			[ $this, 'render_price_display_field', ],
+			[ $this, 'render_price_display_field' ],
 			Settings_Screen::NAME,
 			self::NAME,
 			[
@@ -120,46 +100,54 @@ class Currency extends Settings_Section {
 	}
 
 	public function render_section( $section ) {
-		/**
-		 * This filter is documented in src/BigCommerce/Container/Currency.php
-		 */
-		$auto_format = apply_filters( 'bigcommerce/settings/currency/auto-format', class_exists( '\NumberFormatter' ) );
-
-		if ( ! $auto_format ) {
-			$message = sprintf(
-				__( 'These settings affect how your prices are displayed in WordPress, but will not affect the <a href="%s"> BigCommerce currency settings</a> with which shoppers will be charged.', 'bigcommerce' ),
-				esc_url( 'https://login.bigcommerce.com/deep-links/settings/currencies' )
-			);
-			printf( '<p class="description">%s</p>', $message );
-		}
-
-		$code = get_option( self::CURRENCY_CODE, '' );
-		if ( ! empty( $code ) ) {
-			printf(
-				'<p>%s</p>',
-				sprintf( __( 'Currency code set to %s', 'bigcommerce' ), $code )
-			);
-		} elseif ( $auto_format ) {
-			printf( '<p>%s</p>', __( 'Currency code will be automatically set when the product import completes', 'bigcommerce' ) );
-		}
+		$message = sprintf(
+			__( 'Available currencies are configured in your <a href="%s"> BigCommerce currency settings</a>.', 'bigcommerce' ),
+			esc_url( 'https://login.bigcommerce.com/deep-links/settings/currencies' )
+		);
+		printf( '<p class="description">%s</p>', $message );
 
 		do_action( 'bigcommerce/settings/render/currency', $section );
 	}
 
-	public function render_position_select() {
-		$value   = get_option( self::CURRENCY_SYMBOL_POSITION, 'left' );
-		$choices = [
-			self::POSITION_LEFT        => __( 'Left', 'bigcommerce' ),
-			self::POSITION_RIGHT       => __( 'Right', 'bigcommerce' ),
-			self::POSITION_LEFT_SPACE  => __( 'Left with space', 'bigcommerce' ),
-			self::POSITION_RIGHT_SPACE => __( 'Right with space', 'bigcommerce' ),
-		];
-		$options = [];
-		foreach ( $choices as $key => $label ) {
-			$options[] = sprintf( '<option value="%s" %s>%s</option>', esc_attr( $key ), selected( $key, $value, false ), esc_html( $label ) );
+	public function render_currency_select( $args ) {
+		$channel_term_id = $args['channel'];
+		$currencies      = get_option( self::ENABLED_CURRENCIES, [] );
+		$default         = get_option( self::CURRENCY_CODE, '' );
+
+		$selected = get_term_meta( $channel_term_id, self::CHANNEL_CURRENCY_CODE, true ) ?: $default;
+
+		if ( ! array_key_exists( $selected, $currencies ) ) {
+			if ( $args[ 'count' ] > 1 ) {
+				$message = sprintf( __( 'The currency <strong>%s</strong> is no longer active for your account. Using <strong>%s</strong> for channel <strong>%s</strong>.', 'bigcommerce' ), $selected, $default, get_term( $channel_term_id, Channel\Channel::NAME )->name );
+			} else {
+				$message = sprintf( __( 'The currency <strong>%s</strong> is no longer active for your account. Using <strong>%s</strong> instead.', 'bigcommerce' ), $selected, $default );
+			}
+			printf( '<div class="notice error"><p>%s</p></div>', $message );
 		}
 
-		printf( '<select id="field-%s" name="%s" class="regular-text bc-field-choices">%s</select>', esc_attr( self::CURRENCY_SYMBOL_POSITION ), esc_attr( self::CURRENCY_SYMBOL_POSITION ), implode( "\n", $options ) );
+		if ( empty( $currencies ) ) {
+			$this->render_default_currency_symbol();
+
+			return;
+		}
+
+		if ( count( $currencies ) === 1 ) {
+			echo esc_html( reset( $currencies )['currency_code'] );
+
+			return;
+		}
+
+		printf( '<select name="%1$s[%2$d]" id="field-%1$s-%2$d" class="regular-text bc-field-choices">', esc_attr( self::CHANNEL_CURRENCY_CODE ), $channel_term_id );
+		foreach ( $currencies as $currency ) {
+			printf( '<option value="%1$s" %2$s>%1$s</option>', esc_attr( $currency['currency_code'] ), selected( $selected, $currency['currency_code'], false ) );
+		}
+		echo '</select>';
+
+		if ( $args['count'] > 1 ) {
+			printf( '<p class="description">%s</p>', __( 'Select the currency to use when viewing the store in this channel.', 'bigcommerce' ) );
+		} else {
+			printf( '<p class="description">%s</p>', __( 'Select the currency to use for your store.', 'bigcommerce' ) );
+		}
 	}
 
 	public function render_price_display_field() {
@@ -176,6 +164,35 @@ class Currency extends Settings_Section {
 		printf( '<select id="field-%s" name="%s" class="regular-text bc-field-choices">%s</select>', esc_attr( self::PRICE_DISPLAY ), esc_attr( self::PRICE_DISPLAY ), implode( "\n", $options ) );
 
 		printf( '<p class="description">%s</p>', __( 'Choose whether to include tax in prices shown on your store.', 'bigcommerce' ) );
+	}
+
+	public function render_default_currency_symbol() {
+		$default = get_option( self::CURRENCY_CODE, '' );
+		if ( empty( $default ) ) {
+			printf( '<p>%s</p>', __( 'Currency code will be automatically set when the product import completes', 'bigcommerce' ) );
+		} else {
+			echo esc_html( $default );
+		}
+	}
+
+	/**
+	 * Instead of saving the currency codes for channels as an option,
+	 * save as term meta on each channel
+	 *
+	 * @param array  $new_value
+	 *
+	 * @return bool
+	 * @filter sanitize_option_ . self::CHANNEL_CURRENCY_CODE
+	 */
+	public function save_channel_currencies( $new_value ) {
+		if ( ! is_array( $new_value ) ) {
+			return false;
+		}
+		foreach ( $new_value as $channel_id => $currency_code ) {
+			update_term_meta( $channel_id, self::CHANNEL_CURRENCY_CODE, $currency_code );
+		}
+
+		return false;
 	}
 
 }